requirements.txt
openpyxl
pymysql
selenium
bs4
requests